+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Incorrect reference resolution scope change when resolving responses. [#1184](https://github.com/schemathesis/schemathesis/issues/1184)
 
+### :wrench: Changed
+
+- Display cycles in error messages for infinitely recursive schemas.
+
 ## [4.3.1](https://github.com/schemathesis/schemathesis/compare/v4.3.0...v4.3.1) - 2025-10-14
 
 ### :bug: Fixed